goods/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.0.8
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.3
idna==2.10
isort==5.1.4
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pathspec==0.8.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.26
virtualenv-clone==0.5.4
